Move community reports extractor
