--- graphrag/index/operations/summarize_communities/community_reports_extractor/__init__.py
@@ -1,4 +1,0 @@
-# Copyright (c) 2024 Microsoft Corporation.
-# Licensed under the MIT License
-
-"""The Indexing Engine community reports package root."""
